# openapi.yaml
openapi: 3.1.0
info:
  title: SuteSite API
  description: Personal reference API documentation for sumitsute.com
  version: 1.0.0
  contact:
    name: Sute
servers:
  - url: https://sumitsute.com
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: K
      description: API key (also accepts X-Key header)
    TelegramWebhookAuth:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: SHA256 hash of TELEGRAM_WEBHOOK_SECRET
    BroadcastAuth:
      type: apiKey
      in: header
      name: X-Broadcast-Secret
      description: Must match TELEGRAM_BROADCAST_SECRET
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    BroadcastInput:
      type: object
      properties:
        type:
          type: string
          enum: [bloq]
        title:
          type: string
        slug:
          type: string
        tags:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
      required: [type, title, slug]
    VisitInput:
      type: object
      properties:
        ip:
          type: string
        network:
          type: string
        city:
          type: string
        region:
          type: string
        country_code:
          type: string
        postal:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        org:
          type: string
        timezone:
          type: string
        referrer:
          type: string
    VisitResponse:
      type: object
      properties:
        lastVisitorLocation:
          type: string
          nullable: true
        lastVisitTime:
          type: string
          format: date-time
          nullable: true
        visitorCount:
          type: integer
          nullable: true
paths:
...
  /api/telegram/broadcast:
    post:
      tags: [Telegram]
      summary: Broadcast to Telegram channel
      description: Sends a bloq notification to the configured Telegram channel. Auth uses TELEGRAM_BROADCAST_SECRET.
      security:
        - BroadcastAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BroadcastInput'
      responses:
        '200':
          description: Message broadcasted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  broadcast:
                    type: boolean
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Unsupported broadcast type or invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/visit:
    post:
      tags: [Analytics]
      summary: Track visitor
      description: Records visitor data and returns info about previous visitors.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitInput'
      responses:
        '200':
          description: Visit recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...